GUACAMOLE-268: Use 64-bit unsigned integers to represent file size internally (identical to RDPDR).

+++ src/protocols/rdp/rdp_fs.h
@@ -229,7 +229,7 @@ typedef struct guac_rdp_fs_file {
     /**
      * The size of this file, in bytes.
      */
-    int size;
+    uint64_t size;
 
     /**
      * The time this file was created, as a Windows timestamp.
